requirements.txt
# RB19 Lap Optimizer - MINIMAL (Guaranteed install)
fastf1==3.8.1
streamlit==1.40.0
plotly==5.24.0
pandas==2.2.3
numpy==1.26.4
pyyaml==6.0.2
jupyter==1.0.0
